fourth/distributedpython/requirements.txt
autopep8==2.0.0
numpy==1.23.5
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
tomli==2.0.1
